Fix event persistence error handling
This commit fixes a bug in the error handling logic while attempting to
flush event info objects. Currently, if an error occurs, the exception
is caught and logged. It is not thrown. As a result the processing
continues and the desired events log will not be produced.

The fix throws a custom error corresponding to Events and lets the
caller decide if the processing should continue. As of now the error
should result in termination of the workload or fail unit tests.

Fixes #48