Fix event persistence error handling (#49)
This commit fixes a bug in the error handling logic while attempting to flush event info objects. Currently, if an error occurs while persisting the events, the exception is caught, logged, and ignored. The error is not passed to the caller. As a result the caller continues executing the remaining workflow, even when the benchmark events persistence has failed. Post workflow analysis cannot be done without events.

The fix throws a custom error corresponding to events and lets the caller decide if the processing should continue. As of now the error should result in termination of the workload and cause any unit tests to fail.

Fixes #48