Add degradation metric computation implementation
This change adds a baisc implementation for computing degradataion
metric as described in the paper. The implementation, along with the
helper class, can compute degradation from metric points provided in
input, where each metric point is annotated with the operator (e.g.
phase) it is associated with and the type of metric.

Additionally, tests have been created to ensure the functionality
is working as expected. The project structure has also been updated
to accommodate these new additions.